Adding logging.go file
