--- sdk/azidentity/credential_log_classification.go
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package azidentity
-
-import (
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-)
-
-const (
-	// LogCredential is the log classification that can be used for logging Azure Identity related information
-	LogCredential azcore.LogClassification = "credential"
-)
